# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DK Option Engine
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /execution/preview:
    post:
      operationId: previewExecution
      tags: [execution]
      summary: Build option order legs, compute metrics, and run pre-trade risk validation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExecutionPreviewInput"
      responses:
        "200":
          description: Execution ticket with order legs, metrics, and validation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecutionTicket"
  /execution/submit:
    post:
      operationId: submitExecution
      tags: [execution]
      summary: Submit a validated multi-leg order through the broker (semi-auto only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExecutionSubmitInput"
      responses:
        "201":
          description: Order submitted; pending trade and journal entry created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecutionSubmitResult"
  /execution/adjustment/preview:
    post:
      operationId: previewAdjustmentExecution
      tags: [execution]
      summary: Build a ready-to-submit roll/convert ticket for an open position's recommendation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdjustmentPreviewInput"
      responses:
        "200":
          description: Execution ticket for the recommended roll/convert structure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecutionTicket"
  /execution/adjustment/submit:
    post:
      operationId: submitAdjustmentExecution
      tags: [execution]
      summary: Close the source position and open the recommended roll/convert structure (semi-auto only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdjustmentSubmitInput"
      responses:
        "201":
          description: Source position closed; rolled/converted trade and journal entry created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdjustmentSubmitResult"

  # ─── Full-Auto execution (Phase 6) ──────────────────────────────────────────
...
  /execution/auto/adjust/log:
    get:
      operationId: getAutoAdjustmentLog
      tags: [execution]
      summary: Recent auto-ADJUSTMENT decision audit log (kind=adjust only)
      parameters:
        - name: decision
          in: query
          required: false
          schema:
            type: string
            enum: [executed, skipped, rejected, blocked]
        - name: symbol
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Adjustment audit log entries (newest first)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AutoExecutionLogEntry"
...
components:
  schemas:
...
    ExecutionPreviewInput:
      type: object
      required: [quantity]
      properties:
        scannerResultId:
          type: ["integer", "null"]
        symbol:
          type: ["string", "null"]
        strategy:
          type: ["string", "null"]
          enum: [iron_condor, iron_fly, calendar_spread, earnings, null]
        quantity:
          type: integer
          default: 1
    ExecutionSubmitInput:
      type: object
      required: [quantity, confirm]
      properties:
        scannerResultId:
          type: ["integer", "null"]
        symbol:
          type: ["string", "null"]
        strategy:
          type: ["string", "null"]
          enum: [iron_condor, iron_fly, calendar_spread, earnings, null]
        quantity:
          type: integer
          default: 1
        confirm:
          type: boolean
    ExecutionOrderLeg:
      type: object
      required: [occSymbol, side, optionType, strike, expiration, ratioQty, positionIntent, price]
      properties:
        occSymbol:
          type: string
        side:
          type: string
          enum: [buy, sell]
        optionType:
          type: string
          enum: [call, put]
        strike:
          type: number
        expiration:
          type: string
        ratioQty:
          type: integer
        positionIntent:
          type: string
          enum: [buy_to_open, sell_to_open]
        price:
          type: number
    PreTradeCheck:
      type: object
      required: [label, passed, detail]
      properties:
        label:
          type: string
        passed:
          type: boolean
        detail:
          type: string
    PreTradeValidation:
      type: object
      required: [valid, checks, violations, warnings, riskDollars, riskPct, portfolioRiskBeforePct, portfolioRiskAfterPct]
      properties:
        valid:
          type: boolean
        checks:
          type: array
          items:
            $ref: "#/components/schemas/PreTradeCheck"
        violations:
          type: array
          items:
            type: string
        warnings:
          type: array
          items:
            type: string
        riskDollars:
          type: number
        riskPct:
          type: number
        portfolioRiskBeforePct:
          type: number
        portfolioRiskAfterPct:
          type: number
    ExecutionTicket:
      type: object
      required:
        [symbol, strategy, expiration, daysToExpiry, quantity, legs, netCredit, isCredit, maxProfit, maxLoss, pop, ev, ravishScore, ravishTier, returnOnCapital, buyingPowerRequired, accountValue, riskPct, portfolioRiskBeforePct, portfolioRiskAfterPct, executionMode, canSubmit, validation, warnings]
      properties:
        symbol:
          type: string
        strategy:
          type: string
          enum: [iron_condor, iron_fly, calendar_spread, earnings]
        expiration:
          type: string
        daysToExpiry:
          type: integer
        quantity:
          type: integer
        legs:
          type: array
          items:
            $ref: "#/components/schemas/ExecutionOrderLeg"
        netCredit:
          type: number
        isCredit:
          type: boolean
        maxProfit:
          type: number
        maxLoss:
          type: number
        pop:
          type: number
        ev:
          type: number
        ravishScore:
          type: number
        ravishTier:
          type: string
        returnOnCapital:
          type: number
        buyingPowerRequired:
          type: number
        accountValue:
          type: number
        riskPct:
          type: number
        portfolioRiskBeforePct:
          type: number
        portfolioRiskAfterPct:
          type: number
        executionMode:
          type: string
          enum: [manual, semi_auto, full_auto]
        canSubmit:
          type: boolean
        validation:
          $ref: "#/components/schemas/PreTradeValidation"
        warnings:
          type: array
          items:
            type: string
        adjustment:
          oneOf:
            - $ref: "#/components/schemas/AdjustmentTicketContext"
            - type: "null"
    AdjustmentSourceLeg:
      type: object
      required: [side, optionType, strike, expiration, quantity]
      properties:
        side:
          type: string
          enum: [buy, sell]
        optionType:
          type: string
          enum: [call, put]
        strike:
          type: number
        expiration:
          type: string
        quantity:
          type: integer
    AdjustmentSourcePosition:
      type: object
      required:
        [tradeId, symbol, strategy, strategyLabel, expiration, daysToExpiry, legs, credit, maxProfit, maxLoss, pop, costToClose, currentPnl, isCredit]
      properties:
        tradeId:
          type: integer
        symbol:
          type: string
        strategy:
          type: string
        strategyLabel:
          type: string
        expiration:
          type: ["string", "null"]
        daysToExpiry:
          type: integer
        legs:
          type: array
          items:
            $ref: "#/components/schemas/AdjustmentSourceLeg"
        credit:
          type: number
        maxProfit:
          type: number
        maxLoss:
          type: number
        pop:
          type: number
        costToClose:
          type: number
        currentPnl:
          type: number
        isCredit:
          type: boolean
    AdjustmentTicketContext:
      type: object
      required: [sourceTradeId, action, actionLabel, kind, fromStrategy, toStrategy, closeDescription, rationale, source, netCashflow]
      properties:
        sourceTradeId:
          type: integer
        action:
          type: string
          enum: [roll_threatened, roll_untested, convert]
        actionLabel:
          type: string
        kind:
          type: string
          enum: [roll, convert]
        fromStrategy:
          type: string
        toStrategy:
          type: string
        closeDescription:
          type: string
        rationale:
          type: string
        source:
          $ref: "#/components/schemas/AdjustmentSourcePosition"
        netCashflow:
          type: number
    AdjustmentPreviewInput:
      type: object
      required: [tradeId]
      properties:
        tradeId:
          type: integer
        quantity:
          type: ["integer", "null"]
    AdjustmentSubmitInput:
      type: object
      required: [tradeId, confirm]
      properties:
        tradeId:
          type: integer
        quantity:
          type: ["integer", "null"]
        confirm:
          type: boolean
    AdjustmentSubmitResult:
      type: object
      required: [orderId, status, broker, tradeId, journalId, message, ticket, closedTradeId, closedReason]
      properties:
        orderId:
          type: string
        status:
          type: string
        broker:
          type: string
          enum: [alpaca, mock]
        tradeId:
          type: integer
        journalId:
          type: integer
        message:
          type: string
        ticket:
          $ref: "#/components/schemas/ExecutionTicket"
        closedTradeId:
          type: integer
        closedReason:
          type: string
    ExecutionSubmitResult:
      type: object
      required: [orderId, status, broker, tradeId, journalId, message, ticket]
      properties:
        orderId:
          type: string
        status:
          type: string
        broker:
          type: string
          enum: [alpaca, mock]
        tradeId:
          type: integer
        journalId:
          type: integer
        message:
          type: string
        ticket:
          $ref: "#/components/schemas/ExecutionTicket"

    # ─── Full-Auto execution (Phase 6) ───────────────────────────────────────────
...
    AutoExecutionLogEntry:
      type: object
      required: [id, runId, kind, decision, reason, symbol, strategy, ravishScore, tradeId, alpacaOrderId, createdAt]
      properties:
        id:
          type: integer
        runId:
          type: string
        kind:
          type: string
          enum: [open, adjust]
        decision:
          type: string
          enum: [executed, skipped, rejected, blocked]
        reason:
          type: string
        symbol:
          type: ["string", "null"]
        strategy:
          type: ["string", "null"]
        ravishScore:
          type: number
        tradeId:
          type: ["integer", "null"]
        alpacaOrderId:
          type: ["string", "null"]
        createdAt:
          type: string

    # ─── Backtest ──────────────────────────────────────────────────────────────
...